Implemented DWORD memory access

+++ SadNESs/RAM.h
@@ -6,7 +6,7 @@ class RAM
 public:
 	static void WriteByte(int, byte);
 	static byte ReadByte(int);
-	static void WriteDWORD(int, byte);
+	static void WriteDWORD(int, int);
 	static byte ReadDWORD(int);
 };
